+++ lib/src/analyzers/lint_analyzer/rules/rules_factory.dart
@@ -23,6 +23,7 @@ import 'rules_list/prefer_conditional_expressions/prefer_conditional_expressions
 import 'rules_list/prefer_extracting_callbacks/prefer_extracting_callbacks.dart';
 import 'rules_list/prefer_intl_name/prefer_intl_name.dart';
 import 'rules_list/prefer_on_push_cd_strategy/prefer_on_push_cd_strategy.dart';
+import 'rules_list/prefer_single_widget_per_file/prefer_single_widget_per_file.dart';
 import 'rules_list/prefer_trailing_comma/prefer_trailing_comma.dart';
 import 'rules_list/provide_correct_intl_args/provide_correct_intl_args.dart';
 
@@ -64,6 +65,8 @@ final _implementedRules = <String, Rule Function(Map<String, Object>)>{
   PreferIntlNameRule.ruleId: (config) => PreferIntlNameRule(config),
   PreferOnPushCdStrategyRule.ruleId: (config) =>
       PreferOnPushCdStrategyRule(config),
+  PreferSingleWidgetPerFileRule.ruleId: (config) =>
+      PreferSingleWidgetPerFileRule(config),
   PreferTrailingCommaRule.ruleId: (config) => PreferTrailingCommaRule(config),
   ProvideCorrectIntlArgsRule.ruleId: (config) =>
       ProvideCorrectIntlArgsRule(config),