+++ lib/src/analyzers/lint_analyzer/rules/rules_list/prefer_single_widget_per_file/visitor.dart
@@ -1,0 +1,24 @@
+part of 'prefer_single_widget_per_file.dart';
+
+class _Visitor extends SimpleAstVisitor<void> {
+  final bool _ignorePrivateWidgets;
+
+  final _nodes = <ClassDeclaration>[];
+
+  _Visitor({required bool ignorePrivateWidgets})
+      : _ignorePrivateWidgets = ignorePrivateWidgets;
+
+  Iterable<ClassDeclaration> get nodes =>
+      _nodes.length > 1 ? _nodes.skip(1) : [];
+
+  @override
+  void visitClassDeclaration(ClassDeclaration node) {
+    super.visitClassDeclaration(node);
+
+    final classType = node.extendsClause?.superclass.type;
+    if (isWidgetOrSubclass(classType) &&
+        (!_ignorePrivateWidgets || !Identifier.isPrivateName(node.name.name))) {
+      _nodes.add(node);
+    }
+  }
+}