+++ test/analyzers/lint_analyzer/rules/rules_list/prefer_single_widget_per_file/examples/correct_statefull_widget_example.dart
@@ -1,0 +1,13 @@
+import 'flutter_defines.dart';
+
+class SomeStatefulWidget extends StatefulWidget {
+  @override
+  _someStatefulWidgetState createState() => _someStatefulWidgetState();
+}
+
+class _SomeStatefulWidgetState extends State<InspirationCard> {
+  @override
+  Widget build(BuildContext context) {
+//    ...
+  }
+}