+++ test/analyzers/lint_analyzer/rules/rules_list/prefer_single_widget_per_file/examples/flutter_defines.dart
@@ -1,0 +1,5 @@
+class Widget {}
+
+class StatefulWidget extends Widget {}
+
+class StatelessWidget extends Widget {}