+++ test/analyzers/lint_analyzer/rules/rules_list/prefer_single_widget_per_file/examples/multi_widgets_example.dart
@@ -1,0 +1,27 @@
+import 'flutter_defines.dart';
+
+class ExampleWidget extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+//    ...
+  }
+}
+
+class _PrivateWidget extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+//    ...
+  }
+}
+
+class _AnotherPrivateWidget extends StatefulWidget {
+  @override
+  _SomeStatefulWidgetState createState() => _SomeStatefulWidgetState();
+}
+
+class _SomeStatefulWidgetState extends State<InspirationCard> {
+  @override
+  Widget build(BuildContext context) {
+//    ...
+  }
+}